+++ docs/source/examples/isaaclab/torch_ant_ddpg.py
@@ -2,11 +2,12 @@ import torch
 import torch.nn as nn
 
 # import the skrl components to build the RL system
-from skrl.agents.torch.sac import SAC, SAC_DEFAULT_CONFIG
-from skrl.envs.loaders.torch import load_isaac_orbit_env
+from skrl.agents.torch.ddpg import DDPG, DDPG_DEFAULT_CONFIG
+from skrl.envs.loaders.torch import load_isaaclab_env
 from skrl.envs.wrappers.torch import wrap_env
 from skrl.memories.torch import RandomMemory
-from skrl.models.torch import DeterministicMixin, GaussianMixin, Model
+from skrl.models.torch import DeterministicMixin, Model
+from skrl.resources.noises.torch import OrnsteinUhlenbeckNoise
 from skrl.resources.preprocessors.torch import RunningStandardScaler
 from skrl.trainers.torch import SequentialTrainer
 from skrl.utils import set_seed
@@ -16,12 +17,11 @@ from skrl.utils import set_seed
 set_seed()  # e.g. `set_seed(42)` for fixed seed
 
 
-# define models (stochastic and deterministic models) using mixins
-class StochasticActor(GaussianMixin, Model):
-    def __init__(self, observation_space, action_space, device, clip_actions=False,
-                 clip_log_std=True, min_log_std=-5, max_log_std=2):
+# define models (deterministic models) using mixins
+class DeterministicActor(DeterministicMixin, Model):
+    def __init__(self, observation_space, action_space, device, clip_actions=False):
         Model.__init__(self, observation_space, action_space, device)
-        GaussianMixin.__init__(self, clip_actions, clip_log_std, min_log_std, max_log_std)
+        DeterministicMixin.__init__(self, clip_actions)
 
         self.net = nn.Sequential(nn.Linear(self.num_observations, 512),
                                  nn.ReLU(),
@@ -29,10 +29,9 @@ class StochasticActor(GaussianMixin, Model):
                                  nn.ReLU(),
                                  nn.Linear(256, self.num_actions),
                                  nn.Tanh())
-        self.log_std_parameter = nn.Parameter(torch.zeros(self.num_actions))
 
     def compute(self, inputs, role):
-        return self.net(inputs["states"]), self.log_std_parameter, {}
+        return self.net(inputs["states"]), {}
 
 class Critic(DeterministicMixin, Model):
     def __init__(self, observation_space, action_space, device, clip_actions=False):
@@ -49,8 +48,8 @@ class Critic(DeterministicMixin, Model):
         return self.net(torch.cat([inputs["states"], inputs["taken_actions"]], dim=1)), {}
 
 
-# load and wrap the Isaac Orbit environment
-env = load_isaac_orbit_env(task_name="Isaac-Ant-v0", num_envs=64)
+# load and wrap the Isaac Lab environment
+env = load_isaaclab_env(task_name="Isaac-Ant-v0", num_envs=64)
 env = wrap_env(env)
 
 device = env.device
@@ -61,19 +60,19 @@ memory = RandomMemory(memory_size=15625, num_envs=env.num_envs, device=device)
 
 
 # instantiate the agent's models (function approximators).
-# SAC requires 5 models, visit its documentation for more details
-# https://skrl.readthedocs.io/en/latest/api/agents/sac.html#models
+# DDPG requires 4 models, visit its documentation for more details
+# https://skrl.readthedocs.io/en/latest/api/agents/ddpg.html#models
 models = {}
-models["policy"] = StochasticActor(env.observation_space, env.action_space, device)
-models["critic_1"] = Critic(env.observation_space, env.action_space, device)
-models["critic_2"] = Critic(env.observation_space, env.action_space, device)
-models["target_critic_1"] = Critic(env.observation_space, env.action_space, device)
-models["target_critic_2"] = Critic(env.observation_space, env.action_space, device)
+models["policy"] = DeterministicActor(env.observation_space, env.action_space, device)
+models["target_policy"] = DeterministicActor(env.observation_space, env.action_space, device)
+models["critic"] = Critic(env.observation_space, env.action_space, device)
+models["target_critic"] = Critic(env.observation_space, env.action_space, device)
 
 
 # configure and instantiate the agent (visit its documentation to see all the options)
-# https://skrl.readthedocs.io/en/latest/api/agents/sac.html#configuration-and-hyperparameters
-cfg = SAC_DEFAULT_CONFIG.copy()
+# https://skrl.readthedocs.io/en/latest/api/agents/ddpg.html#configuration-and-hyperparameters
+cfg = DDPG_DEFAULT_CONFIG.copy()
+cfg["exploration"]["noise"] = OrnsteinUhlenbeckNoise(theta=0.15, sigma=0.1, base_scale=0.5, device=device)
 cfg["gradient_steps"] = 1
 cfg["batch_size"] = 4096
 cfg["discount_factor"] = 0.99
@@ -82,10 +81,6 @@ cfg["actor_learning_rate"] = 5e-4
 cfg["critic_learning_rate"] = 5e-4
 cfg["random_timesteps"] = 80
 cfg["learning_starts"] = 80
-cfg["grad_norm_clip"] = 0
-cfg["learn_entropy"] = True
-cfg["entropy_learning_rate"] = 5e-3
-cfg["initial_entropy_value"] = 1.0
 cfg["state_preprocessor"] = RunningStandardScaler
 cfg["state_preprocessor_kwargs"] = {"size": env.observation_space, "device": device}
 # logging to TensorBoard and write checkpoints (in timesteps)
@@ -93,12 +88,12 @@ cfg["experiment"]["write_interval"] = 800
 cfg["experiment"]["checkpoint_interval"] = 8000
 cfg["experiment"]["directory"] = "runs/torch/Isaac-Ant-v0"
 
-agent = SAC(models=models,
-            memory=memory,
-            cfg=cfg,
-            observation_space=env.observation_space,
-            action_space=env.action_space,
-            device=device)
+agent = DDPG(models=models,
+             memory=memory,
+             cfg=cfg,
+             observation_space=env.observation_space,
+             action_space=env.action_space,
+             device=device)
 
 
 # configure and instantiate the RL trainer