+++ docs/source/examples/isaaclab/torch_velocity_anymal_c_ppo.py
@@ -3,7 +3,7 @@ import torch.nn as nn
 
 # import the skrl components to build the RL system
 from skrl.agents.torch.ppo import PPO, PPO_DEFAULT_CONFIG
-from skrl.envs.loaders.torch import load_isaac_orbit_env
+from skrl.envs.loaders.torch import load_isaaclab_env
 from skrl.envs.wrappers.torch import wrap_env
 from skrl.memories.torch import RandomMemory
 from skrl.models.torch import DeterministicMixin, GaussianMixin, Model
@@ -25,17 +25,17 @@ class Shared(GaussianMixin, DeterministicMixin, Model):
         GaussianMixin.__init__(self, clip_actions, clip_log_std, min_log_std, max_log_std, reduction)
         DeterministicMixin.__init__(self, clip_actions)
 
-        self.net = nn.Sequential(nn.Linear(self.num_observations, 256),
+        self.net = nn.Sequential(nn.Linear(self.num_observations, 128),
                                  nn.ELU(),
-                                 nn.Linear(256, 128),
+                                 nn.Linear(128, 128),
                                  nn.ELU(),
-                                 nn.Linear(128, 64),
+                                 nn.Linear(128, 128),
                                  nn.ELU())
 
-        self.mean_layer = nn.Linear(64, self.num_actions)
-        self.log_std_parameter = nn.Parameter(0.5 * torch.ones(self.num_actions))
+        self.mean_layer = nn.Linear(128, self.num_actions)
+        self.log_std_parameter = nn.Parameter(torch.zeros(self.num_actions))
 
-        self.value_layer = nn.Linear(64, 1)
+        self.value_layer = nn.Linear(128, 1)
 
     def act(self, inputs, role):
         if role == "policy":
@@ -45,20 +45,23 @@ class Shared(GaussianMixin, DeterministicMixin, Model):
 
     def compute(self, inputs, role):
         if role == "policy":
-            return torch.tanh(self.mean_layer(self.net(inputs["states"]))), self.log_std_parameter, {}
+            self._shared_output = self.net(inputs["states"])
+            return self.mean_layer(self._shared_output), self.log_std_parameter, {}
         elif role == "value":
-            return self.value_layer(self.net(inputs["states"])), {}
+            shared_output = self.net(inputs["states"]) if self._shared_output is None else self._shared_output
+            self._shared_output = None
+            return self.value_layer(shared_output), {}
 
 
-# load and wrap the Isaac Orbit environment
-env = load_isaac_orbit_env(task_name="Isaac-Reach-Franka-v0")
+# load and wrap the Isaac Lab environment
+env = load_isaaclab_env(task_name="Isaac-Velocity-Flat-Anymal-C-v0")
 env = wrap_env(env)
 
 device = env.device
 
 
 # instantiate a memory as rollout buffer (any memory can be used for this)
-memory = RandomMemory(memory_size=16, num_envs=env.num_envs, device=device)
+memory = RandomMemory(memory_size=24, num_envs=env.num_envs, device=device)
 
 
 # instantiate the agent's models (function approximators).
@@ -72,12 +75,12 @@ models["value"] = models["policy"]  # same instance: shared model
 # configure and instantiate the agent (visit its documentation to see all the options)
 # https://skrl.readthedocs.io/en/latest/api/agents/ppo.html#configuration-and-hyperparameters
 cfg = PPO_DEFAULT_CONFIG.copy()
-cfg["rollouts"] = 16  # memory_size
-cfg["learning_epochs"] = 8
-cfg["mini_batches"] = 8  # 16 * 2048 / 4096
+cfg["rollouts"] = 24  # memory_size
+cfg["learning_epochs"] = 5
+cfg["mini_batches"] = 4  # 24 * 4096 / 24576
 cfg["discount_factor"] = 0.99
 cfg["lambda"] = 0.95
-cfg["learning_rate"] = 3e-4
+cfg["learning_rate"] = 1e-3
 cfg["learning_rate_scheduler"] = KLAdaptiveRL
 cfg["learning_rate_scheduler_kwargs"] = {"kl_threshold": 0.01}
 cfg["random_timesteps"] = 0
@@ -87,7 +90,7 @@ cfg["ratio_clip"] = 0.2
 cfg["value_clip"] = 0.2
 cfg["clip_predicted_values"] = True
 cfg["entropy_loss_scale"] = 0.0
-cfg["value_loss_scale"] = 2.0
+cfg["value_loss_scale"] = 1.0
 cfg["kl_threshold"] = 0
 cfg["rewards_shaper"] = None
 cfg["time_limit_bootstrap"] = False
@@ -96,9 +99,9 @@ cfg["state_preprocessor_kwargs"] = {"size": env.observation_space, "device": dev
 cfg["value_preprocessor"] = RunningStandardScaler
 cfg["value_preprocessor_kwargs"] = {"size": 1, "device": device}
 # logging to TensorBoard and write checkpoints (in timesteps)
-cfg["experiment"]["write_interval"] = 80
-cfg["experiment"]["checkpoint_interval"] = 800
-cfg["experiment"]["directory"] = "runs/torch/Isaac-Reach-Franka-v0"
+cfg["experiment"]["write_interval"] = 60
+cfg["experiment"]["checkpoint_interval"] = 600
+cfg["experiment"]["directory"] = "runs/torch/Isaac-Velocity-Anymal-C-v0"
 
 agent = PPO(models=models,
             memory=memory,
@@ -109,22 +112,8 @@ agent = PPO(models=models,
 
 
 # configure and instantiate the RL trainer
-cfg_trainer = {"timesteps": 16000, "headless": True}
+cfg_trainer = {"timesteps": 12000, "headless": True}
 trainer = SequentialTrainer(cfg=cfg_trainer, env=env, agents=agent)
 
 # start training
 trainer.train()
-
-
-# # ---------------------------------------------------------
-# # comment the code above: `trainer.train()`, and...
-# # uncomment the following lines to evaluate a trained agent
-# # ---------------------------------------------------------
-# from skrl.utils.huggingface import download_model_from_huggingface
-
-# # download the trained agent's checkpoint from Hugging Face Hub and load it
-# path = download_model_from_huggingface("skrl/IsaacOrbit-Isaac-Reach-Franka-v0-PPO", filename="agent.pt")
-# agent.load(path)
-
-# # start evaluation
-# trainer.eval()