+++ skrl/envs/loaders/jax/isaac_orbit_envs.py
@@ -1,3 +1,0 @@
-# since Isaac Orbit environments are implemented on top of PyTorch, the loader is the same
-
-from skrl.envs.loaders.torch import load_isaac_orbit_env